Only try to delete a path if it is a directory
Update @EricLagerg's fix for #180 to more carefully check the
``errno`` attribute of the caught ``OSError`` exception and
only try to use shutil.rmtree if the OSError was raised
because the file is a directory.

This helps prevent getting other instances where OSError would be
raised, for instance a file permission error, from getting eaten.

Addresses https://github.com/openelections/core/issues/180

+++ openelex/base/cache.py
@@ -1,3 +1,4 @@
+import errno
 import os
 import shutil
 
@@ -35,7 +36,10 @@ class StateCache(object):
         for f in files:
             try:
                 os.remove(os.path.join(self.path, f))
-            except OSError:
+            except OSError as e:
+                if e.errno != errno.EISDIR:
+                    raise
+
                 shutil.rmtree(os.path.join(self.path, f))
 
         remaining = self.list_dir()
